Corrected datatype of rotationAngle from int to float

+++ StoryApp/iOS/StoryTelling/StoryTelling/STModifierImageInstanceRotate.h
@@ -10,7 +10,7 @@
 
 @interface STModifierImageInstanceRotate : STModifierImageInstance
 {
-    int rotationAngle;
+    float rotationAngle;
     BOOL isRotationClockwise;
 }
